Channel namespace change
