+++ src/Messages/Channel/Messenger/InvalidCategoryException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Vonage\Messages\MessageType\Messenger;
+namespace Vonage\Messages\Channel\Messenger;
 
 class InvalidCategoryException extends \Exception
 {
