+++ src/Messages/Channel/Messenger/MessengerAudio.php
@@ -1,28 +1,28 @@
 <?php
 
-namespace Vonage\Messages\MessageType\Messenger;
+namespace Vonage\Messages\Channel\Messenger;
 
-use Vonage\Messages\MessageObjects\ImageObject;
-use Vonage\Messages\MessageType\BaseMessage;
+use Vonage\Messages\MessageObjects\AudioObject;
+use Vonage\Messages\Channel\BaseMessage;
 
-class MessengerImage extends BaseMessage
+class MessengerAudio extends BaseMessage
 {
     use MessengerObjectTrait;
 
     protected string $channel = 'messenger';
-    protected string $subType = BaseMessage::MESSAGES_SUBTYPE_IMAGE;
-    protected ImageObject $image;
+    protected string $subType = BaseMessage::MESSAGES_SUBTYPE_AUDIO;
+    protected AudioObject $audioObject;
 
     public function __construct(
         string $to,
         string $from,
-        ImageObject $image,
+        AudioObject $audioObject,
         ?string $category = null,
         ?string $tag = null
     ) {
         $this->to = $to;
         $this->from = $from;
-        $this->image = $image;
+        $this->audioObject = $audioObject;
         $this->category = $category;
         $this->tag = $tag;
     }
@@ -30,7 +30,7 @@ class MessengerImage extends BaseMessage
     public function toArray(): array
     {
         $returnArray = $this->getBaseMessageUniversalOutputArray();
-        $returnArray['image'] = $this->image->toArray();
+        $returnArray['audio'] = $this->audioObject->toArray();
 
         if ($this->requiresMessengerObject()) {
             $returnArray['messenger'] = $this->getMessengerObject();
