+++ src/Messages/Channel/Messenger/MessengerVideo.php
@@ -1,30 +1,40 @@
 <?php
 
-namespace Vonage\Messages\MessageType\WhatsApp;
+namespace Vonage\Messages\Channel\Messenger;
 
 use Vonage\Messages\MessageObjects\VideoObject;
-use Vonage\Messages\MessageType\BaseMessage;
+use Vonage\Messages\Channel\BaseMessage;
 
-class WhatsAppVideo extends BaseMessage
+class MessengerVideo extends BaseMessage
 {
-    protected string $channel = 'whatsapp';
+    use MessengerObjectTrait;
+
+    protected string $channel = 'messenger';
     protected string $subType = BaseMessage::MESSAGES_SUBTYPE_VIDEO;
     protected VideoObject $videoObject;
 
     public function __construct(
         string $to,
         string $from,
-        VideoObject $videoObject
+        VideoObject $videoObject,
+        ?string $category = null,
+        ?string $tag = null
     ) {
         $this->to = $to;
         $this->from = $from;
         $this->videoObject = $videoObject;
+        $this->category = $category;
+        $this->tag = $tag;
     }
 
     public function toArray(): array
     {
         $returnArray = $this->getBaseMessageUniversalOutputArray();
         $returnArray['video'] = $this->videoObject->toArray();
+
+        if ($this->requiresMessengerObject()) {
+            $returnArray['messenger'] = $this->getMessengerObject();
+        }
 
         return $returnArray;
     }
