+++ src/Messages/Channel/Viber/ViberText.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Vonage\Messages\MessageType\Viber;
+namespace Vonage\Messages\Channel\Viber;
 
 use Vonage\Messages\MessageTraits\TextTrait;
-use Vonage\Messages\MessageType\BaseMessage;
+use Vonage\Messages\Channel\BaseMessage;
 
 class ViberText extends BaseMessage
 {
