+++ src/Messages/Channel/WhatsApp/WhatsAppCustom.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Vonage\Messages\MessageType\WhatsApp;
+namespace Vonage\Messages\Channel\WhatsApp;
 
-use Vonage\Messages\MessageType\BaseMessage;
+use Vonage\Messages\Channel\BaseMessage;
 
 class WhatsAppCustom extends BaseMessage
 {
