+++ src/Messages/Channel/WhatsApp/WhatsAppTemplate.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Vonage\Messages\MessageType\WhatsApp;
+namespace Vonage\Messages\Channel\WhatsApp;
 
 use Vonage\Messages\MessageObjects\FileObject;
 use Vonage\Messages\MessageObjects\TemplateObject;
-use Vonage\Messages\MessageType\BaseMessage;
+use Vonage\Messages\Channel\BaseMessage;
 
 class WhatsAppTemplate extends BaseMessage
 {
