+++ src/Messages/Channel/WhatsApp/WhatsAppVideo.php
@@ -1,14 +1,13 @@
 <?php
 
-namespace Vonage\Messages\MessageType\MMS;
+namespace Vonage\Messages\Channel\WhatsApp;
 
-use Vonage\Messages\MessageObjects\AudioObject;
 use Vonage\Messages\MessageObjects\VideoObject;
-use Vonage\Messages\MessageType\BaseMessage;
+use Vonage\Messages\Channel\BaseMessage;
 
-class MMSVideo extends BaseMessage
+class WhatsAppVideo extends BaseMessage
 {
-    protected string $channel = 'mms';
+    protected string $channel = 'whatsapp';
     protected string $subType = BaseMessage::MESSAGES_SUBTYPE_VIDEO;
     protected VideoObject $videoObject;
