+++ test/Messages/ClientTest.php
@@ -23,25 +23,25 @@ use Vonage\Messages\MessageObjects\ImageObject;
 use Vonage\Messages\MessageObjects\TemplateObject;
 use Vonage\Messages\MessageObjects\VCardObject;
 use Vonage\Messages\MessageObjects\VideoObject;
-use Vonage\Messages\MessageType\Messenger\MessengerAudio;
-use Vonage\Messages\MessageType\Messenger\MessengerFile;
-use Vonage\Messages\MessageType\Messenger\MessengerImage;
-use Vonage\Messages\MessageType\Messenger\MessengerText;
-use Vonage\Messages\MessageType\Messenger\MessengerVideo;
-use Vonage\Messages\MessageType\MMS\MMSAudio;
-use Vonage\Messages\MessageType\MMS\MMSImage;
-use Vonage\Messages\MessageType\MMS\MMSvCard;
-use Vonage\Messages\MessageType\MMS\MMSVideo;
-use Vonage\Messages\MessageType\SMS\SMSText;
-use Vonage\Messages\MessageType\Viber\ViberImage;
-use Vonage\Messages\MessageType\Viber\ViberText;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppAudio;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppCustom;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppFile;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppImage;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppTemplate;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppText;
-use Vonage\Messages\MessageType\WhatsApp\WhatsAppVideo;
+use Vonage\Messages\Channel\Messenger\MessengerAudio;
+use Vonage\Messages\Channel\Messenger\MessengerFile;
+use Vonage\Messages\Channel\Messenger\MessengerImage;
+use Vonage\Messages\Channel\Messenger\MessengerText;
+use Vonage\Messages\Channel\Messenger\MessengerVideo;
+use Vonage\Messages\Channel\MMS\MMSAudio;
+use Vonage\Messages\Channel\MMS\MMSImage;
+use Vonage\Messages\Channel\MMS\MMSvCard;
+use Vonage\Messages\Channel\MMS\MMSVideo;
+use Vonage\Messages\Channel\SMS\SMSText;
+use Vonage\Messages\Channel\Viber\ViberImage;
+use Vonage\Messages\Channel\Viber\ViberText;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppAudio;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppCustom;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppFile;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppImage;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppTemplate;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppText;
+use Vonage\Messages\Channel\WhatsApp\WhatsAppVideo;
 use VonageTest\Psr7AssertionTrait;
 use VonageTest\VonageTestCase;
 use Vonage\Client;
